don't eager require serialport
